Restruturando a camada service e controller e inserindo a  logica do metodo em service
